+++ Sourcery/CodeGenerated/Description.generated.swift
@@ -13,10 +13,10 @@ extension ArrayType {
 extension AssociatedValue {
     override var description: String {
         var string = "\(type(of: self)): "
-        string += "localName = \(self.localName), "
-        string += "externalName = \(self.externalName), "
-        string += "annotations = \(self.annotations), "
-        string += "typeName = \(self.typeName)"
+        string += "localName = \(String(describing: self.localName)), "
+        string += "externalName = \(String(describing: self.externalName)), "
+        string += "typeName = \(String(describing: self.typeName)), "
+        string += "annotations = \(String(describing: self.annotations))"
         return string
     }
 }
@@ -110,7 +110,8 @@ extension MethodParameter {
         string += "argumentLabel = \(String(describing: self.argumentLabel)), "
         string += "name = \(String(describing: self.name)), "
         string += "typeName = \(String(describing: self.typeName)), "
-        string += "typeAttributes = \(String(describing: self.typeAttributes))"
+        string += "typeAttributes = \(String(describing: self.typeAttributes)), "
+        string += "annotations = \(String(describing: self.annotations))"
         return string
     }
 }