+++ Sourcery/Parsing/Utils/AnnotationsParser.swift
@@ -55,7 +55,10 @@ internal struct AnnotationsParser {
     /// - Parameter source: Source to extract annotations for.
     /// - Returns: All annotations associated with given source.
     func from(_ source: [String: SourceKitRepresentable]) -> Annotations {
-        guard let range = Substring.key.range(for: source), let lineInfo = location(fromByteOffset: Int(range.offset)).flatMap({ contents.lineAndCharacter(forCharacterOffset: $0) }) else { return [:] }
+        guard let range = Substring.key.range(for: source),
+            let location = contents.location(fromByteOffset: Int(range.offset)),
+            let lineInfo = contents.lineAndCharacter(forCharacterOffset: location)
+            else { return [:] }
 
         var annotations = Annotations()
         for line in lines[0..<lineInfo.line-1].reversed() {
@@ -129,10 +132,10 @@ internal struct AnnotationsParser {
         } else {
             insideBlock = false
             lowerBound = commentLine.range(of: "sourcery:")?.upperBound
-            if commentLine.hasPrefix("/*") {
-                upperBound = commentLine.range(of: "*/")?.lowerBound
-            } else {
+            if commentLine.hasPrefix("//") {
                 upperBound = commentLine.characters.indices.endIndex
+            } else {
+                upperBound = commentLine.range(of: "*/")?.lowerBound
             }
         }
 
@@ -181,9 +184,13 @@ internal struct AnnotationsParser {
         return annotations
     }
 
+}
+
+extension String {
+
     //! this isn't exposed in SourceKitten so we create our own variant
-    private func location(fromByteOffset byteOffset: Int) -> Int? {
-        let lines = contents.lines()
+    func location(fromByteOffset byteOffset: Int) -> Int? {
+        let lines = self.lines()
         if lines.isEmpty {
             return 0
         }
@@ -204,4 +211,5 @@ internal struct AnnotationsParser {
         let utf16Diff = line.content.utf16.distance(from: line.content.utf16.startIndex, to: endUTF16index)
         return line.range.location + utf16Diff
     }
+
 }